feat: refactored usage of shared services in sever-storage package
