+++ packages/server-kit/src/services/authup/module.ts
@@ -5,10 +5,39 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
+import type { ClientOptions, TokenCreatorOptions } from '@authup/core-http-kit';
 import {
-    Client as AuthupClient,
+    Client, mountClientResponseErrorTokenHook as mountAuthupClientResponseErrorTokenHook,
+    mountClientResponseErrorTokenHook,
 } from '@authup/core-http-kit';
+import { isVaultClientUsable, useVaultClient } from '../vault';
 
 export {
-    AuthupClient,
+    mountAuthupClientResponseErrorTokenHook,
 };
+
+export class AuthupClient extends Client {
+    constructor(options: ClientOptions = {}) {
+        super(options);
+
+        let tokenCreator : TokenCreatorOptions;
+        if (isVaultClientUsable()) {
+            tokenCreator = {
+                type: 'robotInVault',
+                name: 'system',
+                vault: useVaultClient(),
+            };
+        } else {
+            tokenCreator = {
+                type: 'user',
+                name: 'admin',
+                password: 'start123',
+            };
+        }
+
+        mountClientResponseErrorTokenHook(this, {
+            baseURL: options.baseURL,
+            tokenCreator,
+        });
+    }
+}
