--- packages/server-storage/src/config/types.ts
@@ -1,12 +1,0 @@
-/*
- * Copyright (c) 2022-2024.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-import type { Client } from 'redis-extension';
-
-export type Config = {
-    redisDatabase: Client
-};
