+++ packages/server-storage/src/core/index.ts
@@ -5,8 +5,5 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-export * from './authup';
 export * from './minio';
-export * from './redis';
 export * from './utils';
-export * from './vault';
