--- packages/server-storage/src/core/vault/index.ts
@@ -1,8 +1,0 @@
-/*
- * Copyright (c) 2024.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-export * from './module';
